Deprecate use of HTTP HEAD: instead using GET piggy-backing
The general approach now is to run an non-Range GET first (instead of
a separate step using HEAD) to both determine the Content-Length and
also to start downloading bytes in one shot, rather than waiting for a
response to HEAD before beginning any bulk data transfer.

This both eliminates some startup time that can compound when dealing
with smaller GETs and can avoid failure when working against a server
that does not support HTTP HEAD in a situation where htcat would
otherwise work fine.

Remove the benchmark section showing slower times on very small files,
because that caveat has now been removed, as the gap has been reduced
to about one tenth of what it was previously.  My measurements of the
same suggest about a 10ms difference between 'curl' and 'htcat', a
small enough difference that adding a whole table to the README is
probably not very illuminating versus its cognitive load.