Add ETag-change detection
If the HTTP implementation supports ETags, check ETags that are
returned with every request to ensure that the resource is not
changing while it is being downloaded.

If the ETag was present on the initial request and changes in a
subsequent Range-GET request, htcat will now return an error.

If an ETag is not present or is of the "weak" variety, allow downloads
to continue, assuming the best.